+++ src/MeisterProPR.Infrastructure/Repositories/MemoryActivityLogRepository.cs
@@ -83,5 +83,5 @@ public sealed partial class MemoryActivityLogRepository(
     [LoggerMessage(
         Level = LogLevel.Warning,
         Message = "Failed to append MemoryActivityLogEntry for thread {ThreadId} / client {ClientId}")]
-    private static partial void LogAppendFailed(ILogger logger, int threadId, Guid clientId, Exception ex);
+    private static partial void LogAppendFailed(ILogger logger, long threadId, Guid clientId, Exception ex);
 }
