+++ tests/MeisterProPR.Application.Tests/Services/PrCrawlServiceTests.cs
@@ -467,7 +467,7 @@ public sealed class PrCrawlServiceTests
                 Arg.Any<Guid>(),
                 Arg.Any<string>(),
                 Arg.Any<int>(),
-                Arg.Any<int>(),
+                Arg.Any<long>(),
                 Arg.Any<string?>(),
                 Arg.Any<string>(),
                 Arg.Any<string>(),
